Direct Bubble Hierarchy Tree

PortfolioOptimisers.AbstractSimilarityMatrixAlgorithm — Type

PortfolioOptimisers.PMFG_T2s — Function

PMFG_T2s(W::AbstractMatrix{<:Real}, nargout::Integer = 3)

Constructs a Triangulated Maximally Filtered Graph (TMFG) starting from a tetrahedron and recursively inserting vertices inside existing triangles (T2 move) in order to approximate a Maximal Planar Graph with the largest total weight, aka Planar Maximally Filtered Graph (PMFG). All weights are non-negative [1].

Arguments

W: N×N matrix of non-negative weights.

nargout: number of output arguments, the same arguments are always returne, this only controls whether some arguments are empty or not.

Outputs

A: adjacency matrix of the PMFG with weights.

tri: list of triangles (triangular faces).

clique3: list of 3-cliques taht are not triangular faces, all 3-cliques are given by [tri; clique3].

cliques: list of all 4-cliques, if nargout <= 3, this will be returned as an empty array.

cliqueTree: 4-cliques tree structure (adjacency matrix), if nargout <= 4, it is returned as an empty array.